Release 1.1 stability (#37)
1.1 is not backwards compatible with 1.1 - major entrypoint types are renamed from <TYPE> to <TYPE>Client, eg. Users -> UsersClient

Improves stability
* added better error handling to User API requests
* removed an unused json->error function

Improves error message quality

Improves testing stability
* encapsulated all calls in the tests the client, so that only one undertest variable is used
* added missing deletes for various endpoint tests
* removed use of duplicated var names, and broke out tests into individual funcs in the slack endpoint testing
* different users per test in users package
* more test isolation, where vars aren't reused between tests...
* alert deletes in tests are running too quickly for the back end API to process them, now delayed and defered

+++ users/user_create_test.go
@@ -1,24 +1,25 @@
 package users_test
 
 import (
+	"github.com/jonboydell/logzio_client/test_utils"
 	"github.com/jonboydell/logzio_client/users"
 	"github.com/stretchr/testify/assert"
 	"testing"
 )
 
 const (
-	test_username = "test@massive.co"
 	test_fullname = "Test User"
 )
 
 func TestUsers_CreateValidUser(t *testing.T) {
 	underTest, err := setupUsersTest()
+	accountId, erx := test_utils.GetAccountId()
 
-	if assert.NoError(t, err) {
+	if assert.NoError(t, err) && assert.NoError(t, erx) && assert.NotZero(t, accountId) {
 		u := users.User{
-			Username:  test_username,
+			Username:  "testcreateuser@massive.co",
 			Fullname:  test_fullname,
-			AccountId: underTest.AccountId,
+			AccountId: accountId,
 			Roles:     []int32{users.UserTypeUser},
 		}
 
@@ -28,9 +29,9 @@ func TestUsers_CreateValidUser(t *testing.T) {
 			v, err := underTest.GetUser(user.Id)
 
 			if assert.NoError(t, err) && assert.NotNil(t, v) {
-				assert.Equal(t, test_username, v.Username)
+				assert.Equal(t, "testcreateuser@massive.co", v.Username)
 				assert.Equal(t, test_fullname, v.Fullname)
-				assert.Equal(t, underTest.AccountId, v.AccountId)
+				assert.Equal(t, accountId, v.AccountId)
 				assert.True(t, v.Active)
 				assert.Equal(t, user.Id, user.Id)
 			}
@@ -43,48 +44,48 @@ func TestUsers_CreateValidUser(t *testing.T) {
 
 func TestUsers_CreateDeleteDuplicateUser(t *testing.T) {
 	underTest, err := setupUsersTest()
+	accountId, erx := test_utils.GetAccountId()
 
-	if assert.NoError(t, err) {
+	if assert.NoError(t, err) && assert.NoError(t, erx) && assert.NotZero(t, accountId) {
 		u := users.User{
-			Username:  test_username,
+			Username:  "testduplicateuser@massive.co",
 			Fullname:  test_fullname,
-			AccountId: underTest.AccountId,
+			AccountId: accountId,
 			Roles:     []int32{users.UserTypeUser},
 		}
 
 		user, err := underTest.CreateUser(u)
-		assert.NoError(t, err)
-		_, err = underTest.CreateUser(u)
-
-		if assert.Error(t, err) {
-			v, err := underTest.GetUser(user.Id)
-
-			if assert.NoError(t, err) && assert.NotNil(t, v) {
-				assert.Equal(t, test_username, v.Username)
-				assert.Equal(t, test_fullname, v.Fullname)
-				assert.Equal(t, underTest.AccountId, v.AccountId)
-				assert.True(t, v.Active)
-				assert.Equal(t, user.Id, user.Id)
-			}
-
-			err = underTest.DeleteUser(user.Id)
-			assert.NoError(t, err)
+		if assert.NoError(t, err) {
+			_, err = underTest.CreateUser(u)
+			assert.Error(t, err)
 		}
+
+		err = underTest.DeleteUser(user.Id)
 	}
 }
 
 func TestUsers_CreateInvalidUser_Email(t *testing.T) {
 	underTest, err := setupUsersTest()
+	accountId, erx := test_utils.GetAccountId()
 
-	if assert.NoError(t, err) {
+	if assert.NoError(t, err) && assert.NoError(t, erx) && assert.NotZero(t, accountId) {
 		u := users.User{
-			Username:  "SomeTestUser",
+			Username:  "InvalidTestUser",
 			Fullname:  "Test User",
-			AccountId: underTest.AccountId,
+			AccountId: accountId,
 			Roles:     []int32{users.UserTypeUser},
 		}
 
 		_, err := underTest.CreateUser(u)
+		assert.Error(t, err)
+	}
+}
+
+func TestUsers_DeleteNonExistingUser(t *testing.T) {
+	underTest, err := setupUsersTest()
+
+	if assert.NoError(t, err) {
+		err = underTest.DeleteUser(21345)
 		assert.Error(t, err)
 	}
 }
